refactor: types util to new helper directory
